added BiteMe-Common project, working fine by owise
