+++ BiteMe-Client/.classpath
@@ -4,5 +4,6 @@
 	<classpathentry combineaccessrules="false" kind="src" path="/OCSF"/>
 	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER"/>
 	<classpathentry kind="con" path="org.eclipse.jdt.USER_LIBRARY/JavaFX"/>
+	<classpathentry combineaccessrules="false" kind="src" path="/BiteMe-Common"/>
 	<classpathentry kind="output" path="bin"/>
 </classpath>
