+++ BiteMe-Client/bin/.gitignore
@@ -1,4 +1,1 @@
 /client/
-/common/
-/gui/
-/logic/
