+++ BiteMe-Client/src/client/ChatClient.java
@@ -7,8 +7,7 @@ package client;
 import ocsf.client.*;
 import client.*;
 import common.ChatIF;
-import logic.Faculty;
-import logic.Student;
+
 
 import java.io.*;
 import java.util.ArrayList;
