+++ BiteMe-Server/bin/.gitignore
@@ -1,3 +1,1 @@
-/Server/
 /common/
-/gui/
